rename to StackedBarChartPage
