+++ src/pages/index.ts
@@ -2,8 +2,5 @@ export { default as Overview } from "./Overview";
 export { default as Feedback } from "./Feedback";
 export { default as SunburstPage } from "./SunburstPage";
 export { default as TreemapPage } from "./TreeMapPage";
-export { default as VaccinationRate } from "./VaccinationRate";
-export { default as G11StackedBarChart } from "./group11/StackedBarChart";
-export { default as SunburstPage } from "./SunburstPage";
-export { default as TreemapPage } from "./TreeMapPage";
+export { default as StackedBarChartPage } from "./StackedBarChartPage";
 export { default as VaccinationRate } from "./VaccinationRate";
